Prevent Duplicate EventHandler Registration in Event
When the KeePass database is synchronized, the ActiveDocumentSelected event triggers, causing the EventHandler SearchController.TextUpdateHandler to be registered twice. This behavior was leading to the issue reported at https://github.com/iamkarlson/keepass-quicksearch/issues/14.

To address this, the commit removes the EventHandler SearchController.TextUpdateHandler before re-adding it. This change prevents duplicate registrations of the event handler. If the event handler is not registered, this operation has no effect; if it is registered, the commit ensures that it exists only once thereafter.